+++ src/assets/track-blocks/TrackInfo.ts
@@ -4,11 +4,12 @@ import { packDelim } from '../constants/packDelim';
 export interface TrackInfo {
   id: number; // track.id
   favorite: boolean;
-  playedCount: number; // track.played_count (but only for current user!).
+  playedCount: number; // track.played_count
   position: number; // position?.inMilliseconds ?? 0
   lastUpdated: number; // DateTime.now().millisecondsSinceEpoch <-> DateTime.fromMillisecondsSinceEpoch(ms)
   lastPlayed: number; // DateTime.now().millisecondsSinceEpoch <-> DateTime.fromMillisecondsSinceEpoch(ms)
   lastFavorited: number; // DateTime
+  // localPlayedCount: number; // (but only for current user!)
 }
 
 const finalPackDelimReg = new RegExp(packDelim + '+$');
@@ -28,6 +29,7 @@ export function trackInfoFromJsonStr(str: string): TrackInfo | undefined {
       lastUpdated, // Timestamp
       lastPlayed, // Timestamp
       lastFavorited,
+      // localPlayedCount,
     ] = list;
     const data: TrackInfo = {
       // Keep the order!
@@ -38,6 +40,7 @@ export function trackInfoFromJsonStr(str: string): TrackInfo | undefined {
       lastUpdated: lastUpdated ? Number(lastUpdated) * 1000 : 0, // Timestamp
       lastPlayed: lastPlayed ? Number(lastPlayed) * 1000 : 0, // Timestamp
       lastFavorited: lastFavorited ? Number(lastFavorited) * 1000 : 0, // Timestamp
+      // localPlayedCount: localPlayedCount ? Number(localPlayedCount) : 0,
     };
     return data;
   } catch (
@@ -62,6 +65,7 @@ export function trackInfoToJsonStr(trackInfo: TrackInfo) {
     lastUpdated,
     lastPlayed,
     lastFavorited,
+    // localPlayedCount,
   } = trackInfo;
   const list = [
     // Keep the order!
@@ -72,6 +76,7 @@ export function trackInfoToJsonStr(trackInfo: TrackInfo) {
     lastUpdated ? Math.round(lastUpdated / 1000) : undefined, // Timestamp
     lastPlayed ? Math.round(lastPlayed / 1000) : undefined, // Timestamp
     lastFavorited ? Math.round(lastFavorited / 1000) : undefined, // Timestamp
+    // localPlayedCount ? Number(localPlayedCount) : undefined,
   ];
   return list.join(packDelim).replace(finalPackDelimReg, '');
 }