+++ src/assets/track-blocks/tracksPlayer.ts
@@ -61,31 +61,27 @@ function calculateAndUpdateTrackPosition(
 function floatingPlayerUpdate(data: FloatingPlayerUpdateData) {
   const { floatingPlayerState, activePlayerData } = data;
   const { id } = activePlayerData;
-  let trackNode = currentTrackPlayer;
-  if (!trackNode || Number(trackNode.dataset.trackId) !== id) {
-    trackNode = getTrackNode(id);
-  }
-  if (!trackNode) {
-    return;
-  }
-  // const isCurrent = trackNode === currentTrackPlayer;
-  const { position = 0, progress = 0, status } = floatingPlayerState;
-  const { dataset } = trackNode;
-  const timeNode = trackNode.querySelector<HTMLElement>('.time');
-  const timeMs = Math.floor(position * 1000);
-  const timeFormatted = formatDuration(timeMs);
-  requestAnimationFrame(() => {
-    if (status) {
-      dataset.status = status;
-    } else {
-      delete dataset.status;
-    }
-    dataset.position = floatToStr(position);
-    dataset.progress = floatToStr(progress);
-    trackNode.style.setProperty('--progress', dataset.progress);
-    if (timeNode) {
-      timeNode.innerText = timeFormatted;
-    }
+  const trackNodes = getTrackNodes(id);
+  trackNodes.forEach((trackNode) => {
+    // const isCurrent = trackNode === currentTrackPlayer;
+    const { position = 0, progress = 0, status } = floatingPlayerState;
+    const { dataset } = trackNode;
+    const timeNode = trackNode.querySelector<HTMLElement>('.time');
+    const timeMs = Math.floor(position * 1000);
+    const timeFormatted = formatDuration(timeMs);
+    requestAnimationFrame(() => {
+      if (status) {
+        dataset.status = status;
+      } else {
+        delete dataset.status;
+      }
+      dataset.position = floatToStr(position);
+      dataset.progress = floatToStr(progress);
+      trackNode.style.setProperty('--progress', dataset.progress);
+      if (timeNode) {
+        timeNode.innerText = timeFormatted;
+      }
+    });
   });
 }
 
@@ -152,10 +148,9 @@ function tryToPlayNextTrack(currentTrackPlayer: HTMLElement) {
   playTrackNode(nextTrackNode);
 }
 
-function getTrackNode(id: number) {
+function getTrackNodes(id: number) {
   const players = Array.from(allPlayers);
-  const trackNode = players.find((it) => Number(it.dataset.trackId) === id);
-  return trackNode;
+  return players.filter((it) => Number(it.id) === id);
 }
 
 function stopPreviousPlayer() {
@@ -173,6 +168,7 @@ function stopPreviousPlayer() {
 function updateTrackPlayedCount(
   trackNode: HTMLElement,
   playedCount?: number,
+  // localPlayedCount?: number,
   _isCurrent?: boolean,
 ) {
   const { dataset } = trackNode;
@@ -181,7 +177,11 @@ function updateTrackPlayedCount(
   if (!id) {
     throw new Error('No current track id!');
   }
-  const updatedTrackInfo = localTrackInfoDb.updatePlayedCount(id, playedCount);
+  const updatedTrackInfo = localTrackInfoDb.updatePlayedCount(
+    id,
+    playedCount,
+    // localPlayedCount,
+  );
   const { playedCount: updatedPlayedCount } = updatedTrackInfo;
   const strValue = quoteHtmlAttr(String(updatedPlayedCount));
   const valueNode = trackNode.querySelector('#played_count') as HTMLElement;
@@ -199,16 +199,12 @@ function updateTrackPlayedCount(
 }
 
 function updateIncrementCallback(data: FloatingPlayerIncrementData) {
-  const {
-    count,
-    // floatingPlayerState,
-    activePlayerData,
-  } = data;
-  const trackNode = getTrackNode(activePlayerData.id);
-  const isCurrent = trackNode === currentTrackPlayer;
-  if (trackNode) {
-    updateTrackPlayedCount(trackNode, count, isCurrent);
-  }
+  const { count, /* localCount, */ activePlayerData } = data;
+  const trackNodes = getTrackNodes(activePlayerData.id);
+  trackNodes.forEach((trackNode) => {
+    const isCurrent = trackNode === currentTrackPlayer;
+    updateTrackPlayedCount(trackNode, count, /* localCount, */ isCurrent);
+  });
 }
 
 function playTrackNode(trackNode: HTMLElement) {
@@ -217,7 +213,7 @@ function playTrackNode(trackNode: HTMLElement) {
     stopPreviousPlayer();
   }
   const { dataset } = trackNode;
-  const id = Number(dataset.trackId);
+  const id = Number(trackNode.id);
 
   const playingId = floatingPlayer.getActiveTrackId();
   const isFloatingPlaying = floatingPlayer.isPlaying();
@@ -276,17 +272,16 @@ function updateTrackFavoriteInDataset(trackNode: HTMLElement, isFavorite: boolea
 }
 
 function updateSingleFavoriteCallback({ id, favorite }: FloatingPlayerFavoriteData) {
-  const trackNode = getTrackNode(id);
-  if (trackNode) {
-    updateTrackFavoriteInDataset(trackNode, favorite);
-  }
+  allPlayers.forEach((trackNode) => {
+    if (Number(trackNode.id) == id) {
+      updateTrackFavoriteInDataset(trackNode, favorite);
+    }
+  });
 }
 
 function updateFavoritesCallback({ favorites }: FloatingPlayerFavoritesData) {
   allPlayers.forEach((trackNode) => {
-    const { dataset } = trackNode;
-    const { trackId } = dataset;
-    const id = Number(trackId);
+    const id = Number(trackNode.id);
     const isFavorite = favorites.includes(id);
     updateTrackFavoriteInDataset(trackNode, isFavorite);
   });
@@ -411,7 +406,9 @@ function initTrackPlayerNode(trackNode: HTMLElement) {
 
 export function initTracksPlayerWrapper(domNode: HTMLElement = document.body) {
   allPlayers = domNode.querySelectorAll<HTMLElement>('.track-player[data-track-media-url]');
+  // Init found track nodes...
   allPlayers.forEach(initTrackPlayerNode);
+  // Init event handlers...
   floatingPlayer.callbacks.addPlayStartCallback(floatingPlayerPlay);
   floatingPlayer.callbacks.addPlayStopCallback(floatingPlayerStop);
   floatingPlayer.callbacks.addUpdateCallback(floatingPlayerUpdate);